css/css-color/parsing/color-invalid-named-color.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Color Level 4: Parsing and serialization of colors using invalid named color notation</title>
<link rel="help" href="https://drafts.csswg.org/css-color-4/#named-colors">
<link rel="help" href="https://drafts.csswg.org/css-color-4/#resolving-sRGB-values">
<link rel="help" href="https://drafts.csswg.org/css-color-4/#serializing-sRGB-values">
<link rel="author" title="Chris Nardi" href="mailto:csnardi1@gmail.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
tests = [
    ["", "Should not parse invalid keyword"],
    [" /* hey */\n", "Should not parse invalid keyword"],
    ["4", "Should not parse invalid keyword"],
    ["top", "Should not parse invalid keyword"],
    ["current-Color", "Should not parse invalid keyword"],
];

for (const test of tests) {
    test_invalid_value("color", test[0]);
}
</script>
</body>
</html>
